requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
google-generativeai
matplotlib
plotly
pillow
